+++ History.md
@@ -45,7 +45,7 @@
 - Add [#111](https://github.com/cleverbeagle/pup/issues/111). Basic data export functionality.
 - Add [#110](https://github.com/cleverbeagle/pup/issues/110). Ability for users to delete their accounts (and data).
 - Add [#109](https://github.com/cleverbeagle/pup/issues/109). Add redirect back to initial request after login.
-- Fix [#106](https://github.com/cleverbeagle/pup/issues/106). Fix missing `medium` link on `<Index />` component.
+- Fix [#106](https://github.com/cleverbeagle/pup/issues/106). Fix missing `medium` link on `<Marketing />` component.
 - Add [#105](https://github.com/cleverbeagle/pup/issues/105). Add `staging` and `production` commands to `package.json`.
 - Fix [#104](https://github.com/cleverbeagle/pup/issues/104). Fix `onCreateUser` not applying default roles to new users.
 - Add [#49](https://github.com/cleverbeagle/pup/issues/49). Add better handling for schema errors inside of Methods. h/t to [@Bandit](https://github.com/Bandit) for the solution.